Modified Products DB, added filtered get by isService
