+++ src/main/java/com/guarderia/products_ms/repositories/Counter2Repository.java
@@ -1,6 +1,6 @@
 package com.guarderia.products_ms.repositories;
 
-import com.guarderia.products_ms.models.Counter2;
+import com.guarderia.products_ms.config.Counter2;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
