+++ src/main/java/com/guarderia/products_ms/repositories/CounterRepository.java
@@ -1,6 +1,6 @@
 package com.guarderia.products_ms.repositories;
 
-import com.guarderia.products_ms.models.Counter;
+import com.guarderia.products_ms.config.Counter;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
