+++ src/main/java/com/guarderia/products_ms/repositories/ProductsRepository.java
@@ -1,8 +1,11 @@
 package com.guarderia.products_ms.repositories;
 
 import com.guarderia.products_ms.models.Product;
+import com.guarderia.products_ms.models.UserProduct;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ProductsRepository extends MongoRepository <Product, Integer> {
+import java.util.List;
 
+public interface ProductsRepository extends MongoRepository <Product, Integer> {
+    List<Product> findByIsService (Boolean isService);
 }
