+++ src/main/java/com/guarderia/products_ms/services/CounterGeneratorService.java
@@ -1,6 +1,6 @@
 package com.guarderia.products_ms.services;
 
-import com.guarderia.products_ms.models.Counter;
+import com.guarderia.products_ms.config.Counter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoOperations;
 import org.springframework.data.mongodb.core.query.Criteria;
@@ -18,14 +18,6 @@ public class CounterGeneratorService {
     @Autowired
     private MongoOperations mongoOperations;
 
-/*
-    @Autowired
-    public CounterGeneratorService(MongoOperations mongoOperations){
-        this.mongoOperations = mongoOperations;
-    }
-*/
-
-
     public Integer getSequenceNumber(String sequenceName){
         Query query = new Query(Criteria.where("id").is(sequenceName));
         Update update = new Update().inc("seq",1);
@@ -36,14 +28,4 @@ public class CounterGeneratorService {
                         Counter.class);
         return !Objects.isNull(counter) ? counter.getSeq() : 1;
     }
-
-
-/*
-    public Integer generateSequence(String seqName){
-        Counter counter = mongoOperations.findAndModify(query(where("_id").is(seqName)),
-                new Update().inc("seq",1), options().returnNew(true).upsert(true),
-                Counter.class);
-        return !Objects.isNull(counter) ? counter.getSeq() : 1;
-    }
-*/
 }
